+++ crates/nix_rs/src/flake/url/mod.rs
@@ -1,0 +1,4 @@
+pub mod attr;
+mod core;
+
+pub use core::*;