+++ crates/nixci/CHANGELOG.md
@@ -2,7 +2,10 @@
 ## Unreleased
 
 - Port to newer `nix_rs`
+- Use `om.ci` as configuration key
 - tests: Removed, and moved to omnix-cli crate.
+- Fix:
+  - Passing `.#foo` where "foo" is missing now errors out, instead of silently defaulting.
 
 ## [1.0.0](https://github.com/srid/nixci/compare/0.5.0...1.0.0) (2024-07-23)